139 fix user model cached (#145)
* reformat usermodelcached + fix tests + added UserProfile

* action: auto-formatting

* fix test

* continue cleaning + fix some tests

* action: auto-formatting

* removed usermodel from friends

* action: auto-formatting

* delete usermodel

* fix some tests

* action: auto-formatting

* action: auto-formatting

* rerun codeclimat

* changes from the review

---------

+++ app/src/main/java/com/epfl/drawyourpath/database/Database.kt
@@ -9,7 +9,7 @@ import java.util.concurrent.CompletableFuture
 data class UserGoals(
     val paths: Long? = null,
     val distance: Double? = null,
-    val activityTime: Long? = null,
+    val activityTime: Double? = null,
 )
 
 data class UserData(
